Create generateParentheses.cpp
Close < Open saved my line of thinking. Also recursion. I should have figured that out when I saw the constraints were small. Not sure if there are any stack methods in C++ that Python had for .join on a string and it would work. Had to refer to the video originally because while I understood the majority of the problem, it was hard to visualize at first how I'd generate ALL possible parentheses. Interesting problem!