Create binarySearch.cpp
Skipped the car fleet problem in stack. That was too much reading and understanding within a short amount of time. Maybe will watch over the weekend to understand (or presumably whenever I have time). Binary search problem is just to get us started with how it works; everything else will get us accustomed to it. Not a bad starter problem.