Create search2DMatrix.cpp
Didn't actually use the video on this one! (: Had to redo once though to get a clear mind of how I wanted to do it. Took my iPad out and started drawing how I wanted to solve it. One mistake I failed to catch is when I set high.

Originally it was high = matrix.size()-1; That is however, incorrect due to the fact that the amount of vectors in the matrix is NOT equal to the amount  of values within a vector inside matrix.

Fun problem to figure and draw out!