Create kokoEatingBananas.cpp
Couldn't visualize the problem in terms of how to solve it. Had to watch the video for this. Eye opening insight to how binary search is used for it though. Had trouble figuring out the issue because I originally did (piles[p] / k) + 1, but if the value ended up sound like (6/6 = 1) it would add + 1 and return the wrong value. So I had to convert the int to a double and then find the ceiling of that in order to get the right value.