Create bestTimeToBuyAndSellStock.cpp
Sliding Window algorithm. Fairly interesting algorithm, however I'm not sure how I feel the difficulty will be for me because I had to watch the video for this easy problem. I'll take it with a grain of salt and just say it was to understand the algorithm and the other problems should be doable.