Create longestRepeatingCharacterReplacement.cpp
Don't have too much familiarity with hash maps through C++, so iterating through it was difficult. Had to watch a video AND look at a solution to understand the conceptual logic behind this problem. I think Sliding Window algorithm may be difficult because I have to get accustomed to using other methods in the form of sliding window. Will revisit.